Revert "#Created a bunch of new components: navbar, carousel, and property"
