+++ package.json
@@ -3,22 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "express": "^4.16.4",
-    "mongodb": "^3.2.3",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
     "react-router-dom": "^5.0.0",
     "react-scripts": "3.0.0"
   },
   "scripts": {
-    "start-frontend": "react-scripts start",
+    "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "start-db": "mongod",
-    "start-backend": "nodemon ./src/Backend/index.js",
-    "start": "npm-run-all -p build start-backend",
-    "dev": "npm-run-all -p start-db start-backend start-frontend"
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -34,10 +28,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "nodemon": "^1.19.0",
-    "npm-run-all": "^4.1.5"
-  },
-  "proxy": "http://localhost:3001"
+  }
 }
