--- src/Backend/controllers/index.controller.js
@@ -1,5 +1,0 @@
-function indexController(req, res) {
-  return res.json("Welcome to my API");
-}
-
-module.exports = indexController;
