--- src/Backend/index.js
@@ -1,8 +1,0 @@
-const App = require("./app");
-
-async function main() {
-  const app = new App(3001);
-  await app.listen();
-}
-
-main();
