--- src/Backend/routes/index.routes.js
@@ -1,6 +1,0 @@
-const router = require("express").Router();
-const indexCotroller = require("./../controllers/index.controller");
-
-router.route("/").get(indexCotroller);
-
-module.exports = router;
