--- src/Frontend/App.js
@@ -1,22 +1,0 @@
-import React from "react";
-import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
-import NavBar from "./components/NavBar";
-import Home from "./components/Home";
-import "./css/App.css";
-
-function App() {
-  return (
-    <div className="App">
-      <Router>
-        <div className="App">
-          <NavBar />
-          <Switch>
-            <Route exact path="/" component={Home} />
-          </Switch>
-        </div>
-      </Router>
-    </div>
-  );
-}
-
-export default App;
