--- src/Frontend/components/Home.js
@@ -1,17 +1,0 @@
-import React from "react";
-//Components
-import Header from "./Header";
-import Body from "./Body";
-import Footer from "./Footer";
-
-export default class Home extends React.Component {
-  render() {
-    return (
-      <div>
-        <Header />
-        <Body />
-        <Footer />
-      </div>
-    );
-  }
-}
