--- src/Frontend/css/Footer.css
@@ -1,3 +1,0 @@
-.Footer {
-    padding-top: 30px;
-}
