+++ src/components/App.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+import NavBar from './NavBar';
+import Home from './Home';
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <Router>
+        <div className="App">
+          <NavBar />
+          <Switch>
+            <Route exact path="/" component={Home} />
+          </Switch>
+        </div>
+      </Router>
+    </div>
+  );
+}
+
+export default App;
