+++ src/components/Footer.css
@@ -1,0 +1,3 @@
+.Footer {
+    padding-top: 30px;
+}
