+++ src/components/Header.css
@@ -1,0 +1,5 @@
+.card-img {
+    opacity: 0.4;
+    box-shadow: 5px 5px 10px rgba(64, 64, 64, 0.5);
+    /* border:1px solid rgb(64,64,64); */
+}
