+++ src/components/Home.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import Header from './Header';
+import Body from './Body';
+import Footer from './Footer';
+
+export default class Home extends React.Component {
+  render() {
+    return (
+      <div>
+        <Header />
+        <Body />
+        <Footer />
+      </div>
+    );
+  }
+}
