+++ src/index.js
@@ -1,10 +1,10 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import "./Frontend/css/index.css";
-import App from "./Frontend/App";
-import * as serviceWorker from "./Frontend/serviceWorker";
+import React from 'react';
+import ReactDOM from 'react-dom';
+import './index.css';
+import App from './components/App';
+import * as serviceWorker from './serviceWorker';
 
-ReactDOM.render(<App />, document.getElementById("root"));
+ReactDOM.render(<App />, document.getElementById('root'));
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
